+++ Benchmarking.Core/Map/RobotStep.cs
@@ -3,9 +3,9 @@
     public class RobotStep
     {
         public Point Location { get; }
-        public int Visibility { get; }
+        public double Visibility { get; }
         
-        public RobotStep(Point location, int visibility)
+        public RobotStep(Point location, double visibility)
         {
             Location = location;
             Visibility = visibility;
